refactor: extract 12 inline components from Navbar.js into separate files
Navbar.js was 1,082 lines with 20 inline component/helper definitions
before the main component. Extracted into dedicated files:

- navbarHelpers.js — getUserDisplayNames, clearBodyScrollStyles, setBodyScrollStyles
- ThemeToggleButton.jsx
- CursorToggleButton.jsx
- AuthButtons.jsx
- MobileNavLink.jsx
- MobileNavGroup.jsx
- DesktopNavLink.jsx
- DesktopNavGroup.jsx
- UserProfileDropdown.jsx
- MobileUserSection.jsx
- BrandMark.jsx
- MobileDrawer.jsx (includes NavList, MobileDrawerHeader, MobileDrawerFooter)

Navbar.js reduced from 1,082 lines to 386 lines (-64%).

Each component retains identical behavior and all props are passed
unchanged through the same interface.
